QA-13542 : allow onChange event triggering on any field and use it on dependent properties
